Switch cmd argument to u_long. This matches what if_ethersubr.c does and allows the code to compile cleanly on amd64 with clang.
Reviewed by:	rwatson
Approved by:	ed (mentor)

+++ sys/net/if.c
@@ -2382,7 +2382,7 @@ ifioctl(struct socket *so, u_long cmd, caddr_t data, struct thread *td)
 		error = (*ifp->if_ioctl)(ifp, cmd, data);
 #else
 	{
-		int ocmd = cmd;
+		u_long ocmd = cmd;
 
 		switch (cmd) {
