+++ sys/netinet6/ip6_mroute.c
@@ -343,7 +343,7 @@ int X_ip6_mforward(struct ip6_hdr *, struct ifnet *, struct mbuf *);
 int X_ip6_mrouter_done(void);
 int X_ip6_mrouter_set(struct socket *, struct sockopt *);
 int X_ip6_mrouter_get(struct socket *, struct sockopt *);
-int X_mrt6_ioctl(int, caddr_t);
+int X_mrt6_ioctl(u_long, caddr_t);
 
 static void
 pim6_init(void)
@@ -449,7 +449,7 @@ X_ip6_mrouter_get(struct socket *so, struct sockopt *sopt)
  * Handle ioctl commands to obtain information from the cache
  */
 int
-X_mrt6_ioctl(int cmd, caddr_t data)
+X_mrt6_ioctl(u_long cmd, caddr_t data)
 {
 	int ret;
